Return match URLs for specified dates

get_afltables_urls returns a character vector containing match URLs for the specified date range

get_afltables_urls(start_date, end_date = Sys.Date())

Arguments

start_date
character string for start date return to URLs from, in "dmy" or "ymd" format

end_date
optional, character string for end date to return URLS, in "dmy" or "ymd" format

Value

a character vector of match URL's between start_date and end_date

Details

This function returns match URLs for the specified date range. This will typically be used to pass to to scrape_afltables_match to return player match results.

Examples

get_afltables_urls("01/01/2018")

#>   [1] "https://afltables.com/afl/stats/games/2018/031420180322.html"
#>   [2] "https://afltables.com/afl/stats/games/2018/010520180323.html"
#>   [3] "https://afltables.com/afl/stats/games/2018/151920180324.html"
#>   [4] "https://afltables.com/afl/stats/games/2018/081320180324.html"
#>   [5] "https://afltables.com/afl/stats/games/2018/122020180324.html"
#>   [6] "https://afltables.com/afl/stats/games/2018/041020180324.html"
#>   [7] "https://afltables.com/afl/stats/games/2018/072120180325.html"
#>   [8] "https://afltables.com/afl/stats/games/2018/091120180325.html"
#>   [9] "https://afltables.com/afl/stats/games/2018/161820180325.html"
#>  [10] "https://afltables.com/afl/stats/games/2018/011420180329.html"
#>  [11] "https://afltables.com/afl/stats/games/2018/121520180330.html"
#>  [12] "https://afltables.com/afl/stats/games/2018/032020180331.html"
#>  [13] "https://afltables.com/afl/stats/games/2018/042120180331.html"
#>  [14] "https://afltables.com/afl/stats/games/2018/111920180331.html"
#>  [15] "https://afltables.com/afl/stats/games/2018/050820180331.html"
#>  [16] "https://afltables.com/afl/stats/games/2018/071820180401.html"
#>  [17] "https://afltables.com/afl/stats/games/2018/131620180401.html"
#>  [18] "https://afltables.com/afl/stats/games/2018/091020180402.html"
#>  [19] "https://afltables.com/afl/stats/games/2018/030420180406.html"
#>  [20] "https://afltables.com/afl/stats/games/2018/131920180407.html"
#>  [21] "https://afltables.com/afl/stats/games/2018/111220180407.html"
#>  [22] "https://afltables.com/afl/stats/games/2018/082020180407.html"
#>  [23] "https://afltables.com/afl/stats/games/2018/011520180407.html"
#>  [24] "https://afltables.com/afl/stats/games/2018/162120180407.html"
#>  [25] "https://afltables.com/afl/stats/games/2018/101420180408.html"
#>  [26] "https://afltables.com/afl/stats/games/2018/050720180408.html"
#>  [27] "https://afltables.com/afl/stats/games/2018/091820180408.html"
#>  [28] "https://afltables.com/afl/stats/games/2018/010420180413.html"
#>  [29] "https://afltables.com/afl/stats/games/2018/082120180414.html"
#>  [30] "https://afltables.com/afl/stats/games/2018/141920180414.html"
#>  [31] "https://afltables.com/afl/stats/games/2018/071620180414.html"
#>  [32] "https://afltables.com/afl/stats/games/2018/031220180414.html"
#>  [33] "https://afltables.com/afl/stats/games/2018/182020180414.html"
#>  [34] "https://afltables.com/afl/stats/games/2018/051320180415.html"
#>  [35] "https://afltables.com/afl/stats/games/2018/101120180415.html"
#>  [36] "https://afltables.com/afl/stats/games/2018/091520180415.html"
#>  [37] "https://afltables.com/afl/stats/games/2018/011620180420.html"
#>  [38] "https://afltables.com/afl/stats/games/2018/152120180421.html"
#>  [39] "https://afltables.com/afl/stats/games/2018/031820180421.html"
#>  [40] "https://afltables.com/afl/stats/games/2018/091320180421.html"
#>  [41] "https://afltables.com/afl/stats/games/2018/070820180421.html"
#>  [42] "https://afltables.com/afl/stats/games/2018/101220180422.html"
#>  [43] "https://afltables.com/afl/stats/games/2018/192020180422.html"
#>  [44] "https://afltables.com/afl/stats/games/2018/111420180424.html"
#>  [45] "https://afltables.com/afl/stats/games/2018/040520180425.html"
#>  [46] "https://afltables.com/afl/stats/games/2018/030720180427.html"
#>  [47] "https://afltables.com/afl/stats/games/2018/091620180428.html"
#>  [48] "https://afltables.com/afl/stats/games/2018/121320180428.html"
#>  [49] "https://afltables.com/afl/stats/games/2018/192120180428.html"
#>  [50] "https://afltables.com/afl/stats/games/2018/101520180428.html"
#>  [51] "https://afltables.com/afl/stats/games/2018/012020180428.html"
#>  [52] "https://afltables.com/afl/stats/games/2018/051120180429.html"
#>  [53] "https://afltables.com/afl/stats/games/2018/041420180429.html"
#>  [54] "https://afltables.com/afl/stats/games/2018/081820180429.html"
#>  [55] "https://afltables.com/afl/stats/games/2018/092120180504.html"
#>  [56] "https://afltables.com/afl/stats/games/2018/072020180505.html"
#>  [57] "https://afltables.com/afl/stats/games/2018/051020180505.html"
#>  [58] "https://afltables.com/afl/stats/games/2018/131820180505.html"
#>  [59] "https://afltables.com/afl/stats/games/2018/121620180505.html"
#>  [60] "https://afltables.com/afl/stats/games/2018/010320180505.html"
#>  [61] "https://afltables.com/afl/stats/games/2018/081420180506.html"
#>  [62] "https://afltables.com/afl/stats/games/2018/111520180506.html"
#>  [63] "https://afltables.com/afl/stats/games/2018/041920180506.html"
#>  [64] "https://afltables.com/afl/stats/games/2018/101620180511.html"
#>  [65] "https://afltables.com/afl/stats/games/2018/182120180512.html"
#>  [66] "https://afltables.com/afl/stats/games/2018/030520180512.html"
#>  [67] "https://afltables.com/afl/stats/games/2018/112020180512.html"
#>  [68] "https://afltables.com/afl/stats/games/2018/011320180512.html"
#>  [69] "https://afltables.com/afl/stats/games/2018/071920180512.html"
#>  [70] "https://afltables.com/afl/stats/games/2018/081520180512.html"
#>  [71] "https://afltables.com/afl/stats/games/2018/121420180513.html"
#>  [72] "https://afltables.com/afl/stats/games/2018/040920180513.html"
#>  [73] "https://afltables.com/afl/stats/games/2018/010720180518.html"
#>  [74] "https://afltables.com/afl/stats/games/2018/122120180519.html"
#>  [75] "https://afltables.com/afl/stats/games/2018/132020180519.html"
#>  [76] "https://afltables.com/afl/stats/games/2018/050920180519.html"
#>  [77] "https://afltables.com/afl/stats/games/2018/041520180519.html"
#>  [78] "https://afltables.com/afl/stats/games/2018/081620180519.html"
#>  [79] "https://afltables.com/afl/stats/games/2018/031120180520.html"
#>  [80] "https://afltables.com/afl/stats/games/2018/101920180520.html"
#>  [81] "https://afltables.com/afl/stats/games/2018/141820180520.html"
#>  [82] "https://afltables.com/afl/stats/games/2018/040720180525.html"
#>  [83] "https://afltables.com/afl/stats/games/2018/141520180526.html"
#>  [84] "https://afltables.com/afl/stats/games/2018/161920180526.html"
#>  [85] "https://afltables.com/afl/stats/games/2018/030920180526.html"
#>  [86] "https://afltables.com/afl/stats/games/2018/052120180526.html"
#>  [87] "https://afltables.com/afl/stats/games/2018/101820180527.html"
#>  [88] "https://afltables.com/afl/stats/games/2018/011120180527.html"
#>  [89] "https://afltables.com/afl/stats/games/2018/081220180527.html"
#>  [90] "https://afltables.com/afl/stats/games/2018/031620180601.html"
#>  [91] "https://afltables.com/afl/stats/games/2018/071120180602.html"
#>  [92] "https://afltables.com/afl/stats/games/2018/101320180602.html"
#>  [93] "https://afltables.com/afl/stats/games/2018/092020180602.html"
#>  [94] "https://afltables.com/afl/stats/games/2018/051420180602.html"
#>  [95] "https://afltables.com/afl/stats/games/2018/151820180602.html"
#>  [96] "https://afltables.com/afl/stats/games/2018/121920180603.html"
#>  [97] "https://afltables.com/afl/stats/games/2018/040820180603.html"
#>  [98] "https://afltables.com/afl/stats/games/2018/012120180603.html"
#>  [99] "https://afltables.com/afl/stats/games/2018/131420180608.html"
#> [100] "https://afltables.com/afl/stats/games/2018/091220180609.html"
#> [101] "https://afltables.com/afl/stats/games/2018/202120180609.html"
#> [102] "https://afltables.com/afl/stats/games/2018/151620180609.html"
#> [103] "https://afltables.com/afl/stats/games/2018/051920180610.html"
#> [104] "https://afltables.com/afl/stats/games/2018/010820180610.html"
#> [105] "https://afltables.com/afl/stats/games/2018/041120180611.html"
#> [106] "https://afltables.com/afl/stats/games/2018/071320180614.html"
#> [107] "https://afltables.com/afl/stats/games/2018/161820180615.html"
#> [108] "https://afltables.com/afl/stats/games/2018/030820180616.html"
#> [109] "https://afltables.com/afl/stats/games/2018/152020180616.html"
#> [110] "https://afltables.com/afl/stats/games/2018/011020180616.html"
#> [111] "https://afltables.com/afl/stats/games/2018/091420180617.html"
#> [112] "https://afltables.com/afl/stats/games/2018/051820180621.html"
#> [113] "https://afltables.com/afl/stats/games/2018/111320180622.html"
#> [114] "https://afltables.com/afl/stats/games/2018/102020180623.html"
#> [115] "https://afltables.com/afl/stats/games/2018/192120180623.html"
#> [116] "https://afltables.com/afl/stats/games/2018/071220180623.html"
#> [117] "https://afltables.com/afl/stats/games/2018/030420180624.html"
#> [118] "https://afltables.com/afl/stats/games/2018/141620180628.html"
#> [119] "https://afltables.com/afl/stats/games/2018/070920180629.html"
#> [120] "https://afltables.com/afl/stats/games/2018/031320180630.html"
#> [121] "https://afltables.com/afl/stats/games/2018/011820180630.html"
#> [122] "https://afltables.com/afl/stats/games/2018/042020180630.html"
#> [123] "https://afltables.com/afl/stats/games/2018/102120180630.html"
#> [124] "https://afltables.com/afl/stats/games/2018/111520180701.html"
#> [125] "https://afltables.com/afl/stats/games/2018/051220180701.html"
#> [126] "https://afltables.com/afl/stats/games/2018/081920180701.html"
#> [127] "https://afltables.com/afl/stats/games/2018/091620180705.html"
#> [128] "https://afltables.com/afl/stats/games/2018/011420180706.html"
#> [129] "https://afltables.com/afl/stats/games/2018/031920180707.html"
#> [130] "https://afltables.com/afl/stats/games/2018/131520180707.html"
#> [131] "https://afltables.com/afl/stats/games/2018/071020180707.html"
#> [132] "https://afltables.com/afl/stats/games/2018/081120180707.html"
#> [133] "https://afltables.com/afl/stats/games/2018/122020180708.html"
#> [134] "https://afltables.com/afl/stats/games/2018/040520180708.html"
#> [135] "https://afltables.com/afl/stats/games/2018/182120180708.html"
#> [136] "https://afltables.com/afl/stats/games/2018/010920180712.html"
#> [137] "https://afltables.com/afl/stats/games/2018/031520180713.html"
#> [138] "https://afltables.com/afl/stats/games/2018/101920180714.html"
#> [139] "https://afltables.com/afl/stats/games/2018/071120180714.html"
#> [140] "https://afltables.com/afl/stats/games/2018/052020180714.html"
#> [141] "https://afltables.com/afl/stats/games/2018/142120180714.html"
#> [142] "https://afltables.com/afl/stats/games/2018/041820180715.html"
#> [143] "https://afltables.com/afl/stats/games/2018/121620180715.html"
#> [144] "https://afltables.com/afl/stats/games/2018/081320180715.html"
#> [145] "https://afltables.com/afl/stats/games/2018/141520180720.html"
#> [146] "https://afltables.com/afl/stats/games/2018/041220180721.html"
#> [147] "https://afltables.com/afl/stats/games/2018/162020180721.html"
#> [148] "https://afltables.com/afl/stats/games/2018/050820180721.html"
#> [149] "https://afltables.com/afl/stats/games/2018/011920180721.html"
#> [150] "https://afltables.com/afl/stats/games/2018/091120180721.html"
#> [151] "https://afltables.com/afl/stats/games/2018/031020180722.html"
#> [152] "https://afltables.com/afl/stats/games/2018/071820180722.html"
#> [153] "https://afltables.com/afl/stats/games/2018/132120180722.html"
#> [154] "https://afltables.com/afl/stats/games/2018/051620180727.html"
#> [155] "https://afltables.com/afl/stats/games/2018/041420180728.html"
#> [156] "https://afltables.com/afl/stats/games/2018/091920180728.html"
#> [157] "https://afltables.com/afl/stats/games/2018/152120180728.html"
#> [158] "https://afltables.com/afl/stats/games/2018/032020180728.html"
#> [159] "https://afltables.com/afl/stats/games/2018/011120180728.html"
#> [160] "https://afltables.com/afl/stats/games/2018/121820180729.html"
#> [161] "https://afltables.com/afl/stats/games/2018/071320180729.html"
#> [162] "https://afltables.com/afl/stats/games/2018/081020180729.html"
#> [163] "https://afltables.com/afl/stats/games/2018/091420180803.html"
#> [164] "https://afltables.com/afl/stats/games/2018/051020180804.html"
#> [165] "https://afltables.com/afl/stats/games/2018/121920180804.html"
#> [166] "https://afltables.com/afl/stats/games/2018/011320180804.html"
#> [167] "https://afltables.com/afl/stats/games/2018/041620180804.html"
#> [168] "https://afltables.com/afl/stats/games/2018/071520180804.html"
#> [169] "https://afltables.com/afl/stats/games/2018/032120180805.html"
#> [170] "https://afltables.com/afl/stats/games/2018/112020180805.html"
#> [171] "https://afltables.com/afl/stats/games/2018/081820180805.html"
#> [172] "https://afltables.com/afl/stats/games/2018/051520180810.html"
#> [173] "https://afltables.com/afl/stats/games/2018/091020180811.html"
#> [174] "https://afltables.com/afl/stats/games/2018/142020180811.html"
#> [175] "https://afltables.com/afl/stats/games/2018/131820180811.html"
#> [176] "https://afltables.com/afl/stats/games/2018/012120180811.html"
#> [177] "https://afltables.com/afl/stats/games/2018/041920180811.html"
#> [178] "https://afltables.com/afl/stats/games/2018/071220180812.html"
#> [179] "https://afltables.com/afl/stats/games/2018/111620180812.html"
#> [180] "https://afltables.com/afl/stats/games/2018/030820180812.html"
#> [181] "https://afltables.com/afl/stats/games/2018/051420180817.html"
#> [182] "https://afltables.com/afl/stats/games/2018/041320180818.html"
#> [183] "https://afltables.com/afl/stats/games/2018/080920180818.html"
#> [184] "https://afltables.com/afl/stats/games/2018/162120180818.html"
#> [185] "https://afltables.com/afl/stats/games/2018/101520180818.html"
#> [186] "https://afltables.com/afl/stats/games/2018/192020180818.html"
#> [187] "https://afltables.com/afl/stats/games/2018/030720180819.html"
#> [188] "https://afltables.com/afl/stats/games/2018/111820180819.html"
#> [189] "https://afltables.com/afl/stats/games/2018/011220180819.html"
#> [190] "https://afltables.com/afl/stats/games/2018/051320180824.html"
#> [191] "https://afltables.com/afl/stats/games/2018/092020180825.html"
#> [192] "https://afltables.com/afl/stats/games/2018/071420180825.html"
#> [193] "https://afltables.com/afl/stats/games/2018/040820180825.html"
#> [194] "https://afltables.com/afl/stats/games/2018/010320180825.html"
#> [195] "https://afltables.com/afl/stats/games/2018/101620180825.html"
#> [196] "https://afltables.com/afl/stats/games/2018/181920180826.html"
#> [197] "https://afltables.com/afl/stats/games/2018/112120180826.html"
#> [198] "https://afltables.com/afl/stats/games/2018/121520180826.html"
#> [199] "https://afltables.com/afl/stats/games/2018/101420180906.html"
#> [200] "https://afltables.com/afl/stats/games/2018/091120180907.html"
#> [201] "https://afltables.com/afl/stats/games/2018/162120180908.html"
#> [202] "https://afltables.com/afl/stats/games/2018/041820180908.html"
#> [203] "https://afltables.com/afl/stats/games/2018/101120180914.html"
#> [204] "https://afltables.com/afl/stats/games/2018/042120180915.html"
#> [205] "https://afltables.com/afl/stats/games/2018/041420180921.html"
#> [206] "https://afltables.com/afl/stats/games/2018/111820180922.html"
#> [207] "https://afltables.com/afl/stats/games/2018/041820180929.html"

get_afltables_urls("01/01/2018", end_date = "01/04/2018")

#>  [1] "https://afltables.com/afl/stats/games/2018/031420180322.html"
#>  [2] "https://afltables.com/afl/stats/games/2018/010520180323.html"
#>  [3] "https://afltables.com/afl/stats/games/2018/151920180324.html"
#>  [4] "https://afltables.com/afl/stats/games/2018/081320180324.html"
#>  [5] "https://afltables.com/afl/stats/games/2018/122020180324.html"
#>  [6] "https://afltables.com/afl/stats/games/2018/041020180324.html"
#>  [7] "https://afltables.com/afl/stats/games/2018/072120180325.html"
#>  [8] "https://afltables.com/afl/stats/games/2018/091120180325.html"
#>  [9] "https://afltables.com/afl/stats/games/2018/161820180325.html"
#> [10] "https://afltables.com/afl/stats/games/2018/011420180329.html"
#> [11] "https://afltables.com/afl/stats/games/2018/121520180330.html"
#> [12] "https://afltables.com/afl/stats/games/2018/032020180331.html"
#> [13] "https://afltables.com/afl/stats/games/2018/042120180331.html"
#> [14] "https://afltables.com/afl/stats/games/2018/111920180331.html"
#> [15] "https://afltables.com/afl/stats/games/2018/050820180331.html"