Get AFLW match data

Retrieves AFLW match data for all available matches. Sources data from http://www.afl.com.au/womens/matches/stats

get_aflw_match_data(start_year = 2017)

Arguments

start_year
optional, integer for start year to return match data onwards from

Value

a data frame of data for all available AFLW matches

Examples

## All data
get_aflw_match_data()

#> Warning: Scores not present for round CD_R201926401, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926402, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926403, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926404, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926405, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926406, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926407, returning match fixture information only.

#> # A tibble: 93 x 30
#>    Match.Id Round.Id Competition.Id Venue Local.Start.Time    Round.Number
#>    <chr>    <chr>    <chr>          <chr> <dttm>                     <int>
#>  1 CD_M201… CD_R201… CD_S2017264    Ikon… 2017-02-03 19:45:00            1
#>  2 CD_M201… CD_R201… CD_S2017264    Theb… 2017-02-04 16:35:00            1
#>  3 CD_M201… CD_R201… CD_S2017264    VU W… 2017-02-04 19:40:00            1
#>  4 CD_M201… CD_R201… CD_S2017264    Case… 2017-02-05 17:05:00            1
#>  5 CD_M201… CD_R201… CD_S2017264    VU W… 2017-02-10 19:35:00            2
#>  6 CD_M201… CD_R201… CD_S2017264    Ikon… 2017-02-11 15:35:00            2
#>  7 CD_M201… CD_R201… CD_S2017264    Ikon… 2017-02-11 19:40:00            2
#>  8 CD_M201… CD_R201… CD_S2017264    Frem… 2017-02-12 16:05:00            2
#>  9 CD_M201… CD_R201… CD_S2017264    Blac… 2017-02-18 12:05:00            3
#> 10 CD_M201… CD_R201… CD_S2017264    Sout… 2017-02-18 15:35:00            3
#> # … with 83 more rows, and 24 more variables: Round.Abbreviation <chr>,
#> #   Weather.Type <chr>, Weather.Description <chr>, Temperature <dbl>,
#> #   Home.Team <chr>, Home.Goals <int>, Home.Behinds <int>, Home.Points <int>,
#> #   Home.Left.Behinds <int>, Home.Right.Behinds <int>, Home.Left.Posters <int>,
#> #   Home.Right.Posters <int>, Home.Rushed.Behinds <int>,
#> #   Home.Touched.Behinds <int>, Away.Team <chr>, Away.Goals <int>,
#> #   Away.Behinds <int>, Away.Points <int>, Away.Left.Behinds <int>,
#> #   Away.Right.Behinds <int>, Away.Left.Posters <int>,
#> #   Away.Right.Posters <int>, Away.Rushed.Behinds <int>,
#> #   Away.Touched.Behinds <int>

## 2018 data onward
get_aflw_match_data(start_year = 2018)

#> Warning: Scores not present for round CD_R201926401, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926402, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926403, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926404, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926405, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926406, returning match fixture information only.

#> Warning: Scores not present for round CD_R201926407, returning match fixture information only.

#> # A tibble: 64 x 30
#>    Match.Id Round.Id Competition.Id Venue Local.Start.Time    Round.Number
#>    <chr>    <chr>    <chr>          <chr> <dttm>                     <int>
#>  1 CD_M201… CD_R201… CD_S2018264    Ikon… 2018-02-02 19:45:00            1
#>  2 CD_M201… CD_R201… CD_S2018264    Case… 2018-02-03 17:05:00            1
#>  3 CD_M201… CD_R201… CD_S2018264    Norw… 2018-02-03 18:45:00            1
#>  4 CD_M201… CD_R201… CD_S2018264    VU W… 2018-02-04 16:35:00            1
#>  5 CD_M201… CD_R201… CD_S2018264    Drum… 2018-02-09 19:05:00            2
#>  6 CD_M201… CD_R201… CD_S2018264    Case… 2018-02-10 17:05:00            2
#>  7 CD_M201… CD_R201… CD_S2018264    Optu… 2018-02-10 16:10:00            2
#>  8 CD_M201… CD_R201… CD_S2018264    Sout… 2018-02-11 15:35:00            2
#>  9 CD_M201… CD_R201… CD_S2018264    Norw… 2018-02-17 13:35:00            3
#> 10 CD_M201… CD_R201… CD_S2018264    Ikon… 2018-02-17 19:45:00            3
#> # … with 54 more rows, and 24 more variables: Round.Abbreviation <chr>,
#> #   Weather.Type <chr>, Weather.Description <chr>, Temperature <dbl>,
#> #   Home.Team <chr>, Home.Goals <int>, Home.Behinds <int>, Home.Points <int>,
#> #   Home.Left.Behinds <int>, Home.Right.Behinds <int>, Home.Left.Posters <int>,
#> #   Home.Right.Posters <int>, Home.Rushed.Behinds <int>,
#> #   Home.Touched.Behinds <int>, Away.Team <chr>, Away.Goals <int>,
#> #   Away.Behinds <int>, Away.Points <int>, Away.Left.Behinds <int>,
#> #   Away.Right.Behinds <int>, Away.Left.Posters <int>,
#> #   Away.Right.Posters <int>, Away.Rushed.Behinds <int>,
#> #   Away.Touched.Behinds <int>