Get basic match results from afltables.com

get_match_results returns a dataframe containing all match results from 1897-current

get_match_results()

Value

Returns a data frame containing a line for each match

Details

The dataframe contains information about the Date, teams involved, scores and venue. It comes from afltables 'big lists' section. This is a limited dataset but is very fast to access. It generally is updated on the day after the last game

Examples

# \donttest{
get_match_results()

#> # A tibble: 15,614 x 16
#>     Game Date       Round Home.Team Home.Goals Home.Behinds Home.Points
#>    <dbl> <date>     <chr> <chr>          <int>        <int>       <int>
#>  1     1 1897-05-08 R1    Fitzroy            6           13          49
#>  2     2 1897-05-08 R1    Collingw…          5           11          41
#>  3     3 1897-05-08 R1    Geelong            3            6          24
#>  4     4 1897-05-08 R1    Sydney             3            9          27
#>  5     5 1897-05-15 R2    Sydney             6            4          40
#>  6     6 1897-05-15 R2    Essendon           4            6          30
#>  7     7 1897-05-15 R2    St Kilda           3            8          26
#>  8     8 1897-05-15 R2    Melbourne          9           10          64
#>  9     9 1897-05-22 R3    Collingw…          6            5          41
#> 10    10 1897-05-22 R3    Fitzroy            5            9          39
#> # … with 15,604 more rows, and 9 more variables: Away.Team <chr>,
#> #   Away.Goals <int>, Away.Behinds <int>, Away.Points <int>, Venue <chr>,
#> #   Margin <int>, Season <dbl>, Round.Type <chr>, Round.Number <int>

# }