Recreate the ladder for every or any given round and/or season

return_ladder returns a dataframe containing the ladder for either all seasons and rounds since 1987, or individual rounds/seasons

return_ladder(match_results_df = NA, season_round = NA, season = NA)

Arguments

match_results_df
A dataframe that has been returned from get_match_results. If empty get_match_results will execute first

season_round
An integer of the round or vector of integers for multiple rounds. If empty, all rounds returned

season
An integer of the season or vector of integers for multiple seasons. If empty, all seasons returned

Value

Returns a data frame containing a line for each team's ladder position at each round of a season

Details

The dataframe contains information about the Round, Season, Points For/Against, Ladder Position. It can either take in a data frame created using get_match_results, or if match_results_df is unspecified, will extract all games using get_match_results. Will only allow selecting rounds of the premiership season, not finals.

Examples

if (FALSE) {
return_ladder()
return_ladder(match_results_df=get_match_results_df, season_round=23, season=1990:2019)
return_ladder(season_round = 10, season = 2019)
}