Return afltables player match stats

scrape_afltables_match returns a character vector containing match URLs for the specified date range

scrape_afltables_match(match_urls)

Arguments

match_urls
A list of URL's for matches to scrape data from

Value

data table of afltables.com match results, with a row per player per match.

Details

This function returns the full afltables.com match stats for each player and each game specified in match_urls. It is useful to use the helper function get_afltables_urls to return these or simply navigate to afltables.com and find the match of interest.

Examples

# \donttest{
scrape_afltables_match("https://afltables.com/afl/stats/games/2018/071120180602.html")

#> Downloading data

#> 
#> Finished downloading data. Processing XMLs

#>    Season Round       Date Local.start.time      Venue Attendance
#> 1    2018    11 2018-06-02             1345 Docklands       28485
#> 2    2018    11 2018-06-02             1345 Docklands       28485
#> 3    2018    11 2018-06-02             1345 Docklands       28485
#> 4    2018    11 2018-06-02             1345 Docklands       28485
#> 5    2018    11 2018-06-02             1345 Docklands       28485
#> 6    2018    11 2018-06-02             1345 Docklands       28485
#> 7    2018    11 2018-06-02             1345 Docklands       28485
#> 8    2018    11 2018-06-02             1345 Docklands       28485
#> 9    2018    11 2018-06-02             1345 Docklands       28485
#> 10   2018    11 2018-06-02             1345 Docklands       28485
#> 11   2018    11 2018-06-02             1345 Docklands       28485
#> 12   2018    11 2018-06-02             1345 Docklands       28485
#> 13   2018    11 2018-06-02             1345 Docklands       28485
#> 14   2018    11 2018-06-02             1345 Docklands       28485
#> 15   2018    11 2018-06-02             1345 Docklands       28485
#> 16   2018    11 2018-06-02             1345 Docklands       28485
#> 17   2018    11 2018-06-02             1345 Docklands       28485
#> 18   2018    11 2018-06-02             1345 Docklands       28485
#> 19   2018    11 2018-06-02             1345 Docklands       28485
#> 20   2018    11 2018-06-02             1345 Docklands       28485
#> 21   2018    11 2018-06-02             1345 Docklands       28485
#> 22   2018    11 2018-06-02             1345 Docklands       28485
#> 23   2018    11 2018-06-02             1345 Docklands       28485
#> 24   2018    11 2018-06-02             1345 Docklands       28485
#> 25   2018    11 2018-06-02             1345 Docklands       28485
#> 26   2018    11 2018-06-02             1345 Docklands       28485
#> 27   2018    11 2018-06-02             1345 Docklands       28485
#> 28   2018    11 2018-06-02             1345 Docklands       28485
#> 29   2018    11 2018-06-02             1345 Docklands       28485
#> 30   2018    11 2018-06-02             1345 Docklands       28485
#> 31   2018    11 2018-06-02             1345 Docklands       28485
#> 32   2018    11 2018-06-02             1345 Docklands       28485
#> 33   2018    11 2018-06-02             1345 Docklands       28485
#> 34   2018    11 2018-06-02             1345 Docklands       28485
#> 35   2018    11 2018-06-02             1345 Docklands       28485
#> 36   2018    11 2018-06-02             1345 Docklands       28485
#> 37   2018    11 2018-06-02             1345 Docklands       28485
#> 38   2018    11 2018-06-02             1345 Docklands       28485
#> 39   2018    11 2018-06-02             1345 Docklands       28485
#> 40   2018    11 2018-06-02             1345 Docklands       28485
#> 41   2018    11 2018-06-02             1345 Docklands       28485
#> 42   2018    11 2018-06-02             1345 Docklands       28485
#> 43   2018    11 2018-06-02             1345 Docklands       28485
#> 44   2018    11 2018-06-02             1345 Docklands       28485
#>           Home.team HQ1G HQ1B HQ2G HQ2B HQ3G HQ3B HQ4G HQ4B Home.score
#> 1  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 2  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 3  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 4  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 5  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 6  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 7  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 8  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 9  Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 10 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 11 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 12 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 13 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 14 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 15 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 16 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 17 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 18 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 19 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 20 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 21 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 22 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 23 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 24 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 25 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 26 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 27 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 28 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 29 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 30 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 31 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 32 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 33 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 34 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 35 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 36 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 37 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 38 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 39 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 40 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 41 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 42 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 43 Western Bulldogs    3    3    4    6    6    8    7    9         51
#> 44 Western Bulldogs    3    3    4    6    6    8    7    9         51
#>    Away.team AQ1G AQ1B AQ2G AQ2B AQ3G AQ3B AQ4G AQ4B Away.score First.name
#> 1  Melbourne    2    1    7    4   11    7   15   10        100     Marcus
#> 2  Melbourne    2    1    7    4   11    7   15   10        100        Tom
#> 3  Melbourne    2    1    7    4   11    7   15   10        100     Hayden
#> 4  Melbourne    2    1    7    4   11    7   15   10        100       Luke
#> 5  Melbourne    2    1    7    4   11    7   15   10        100      Caleb
#> 6  Melbourne    2    1    7    4   11    7   15   10        100       Tory
#> 7  Melbourne    2    1    7    4   11    7   15   10        100      Billy
#> 8  Melbourne    2    1    7    4   11    7   15   10        100      Mitch
#> 9  Melbourne    2    1    7    4   11    7   15   10        100     Lachie
#> 10 Melbourne    2    1    7    4   11    7   15   10        100      Jason
#> 11 Melbourne    2    1    7    4   11    7   15   10        100    Patrick
#> 12 Melbourne    2    1    7    4   11    7   15   10        100       Jack
#> 13 Melbourne    2    1    7    4   11    7   15   10        100       Toby
#> 14 Melbourne    2    1    7    4   11    7   15   10        100       Dale
#> 15 Melbourne    2    1    7    4   11    7   15   10        100         Ed
#> 16 Melbourne    2    1    7    4   11    7   15   10        100     Jordan
#> 17 Melbourne    2    1    7    4   11    7   15   10        100       Josh
#> 18 Melbourne    2    1    7    4   11    7   15   10        100     Roarke
#> 19 Melbourne    2    1    7    4   11    7   15   10        100       Matt
#> 20 Melbourne    2    1    7    4   11    7   15   10        100    Jackson
#> 21 Melbourne    2    1    7    4   11    7   15   10        100     Bailey
#> 22 Melbourne    2    1    7    4   11    7   15   10        100     Easton
#> 23 Melbourne    2    1    7    4   11    7   15   10        100      Angus
#> 24 Melbourne    2    1    7    4   11    7   15   10        100     Bayley
#> 25 Melbourne    2    1    7    4   11    7   15   10        100        Max
#> 26 Melbourne    2    1    7    4   11    7   15   10        100      Mitch
#> 27 Melbourne    2    1    7    4   11    7   15   10        100      James
#> 28 Melbourne    2    1    7    4   11    7   15   10        100    Michael
#> 29 Melbourne    2    1    7    4   11    7   15   10        100      Jesse
#> 30 Melbourne    2    1    7    4   11    7   15   10        100    Neville
#> 31 Melbourne    2    1    7    4   11    7   15   10        100     Nathan
#> 32 Melbourne    2    1    7    4   11    7   15   10        100       Jake
#> 33 Melbourne    2    1    7    4   11    7   15   10        100     Jordan
#> 34 Melbourne    2    1    7    4   11    7   15   10        100      Oscar
#> 35 Melbourne    2    1    7    4   11    7   15   10        100        Tom
#> 36 Melbourne    2    1    7    4   11    7   15   10        100       Jake
#> 37 Melbourne    2    1    7    4   11    7   15   10        100       Alex
#> 38 Melbourne    2    1    7    4   11    7   15   10        100    Clayton
#> 39 Melbourne    2    1    7    4   11    7   15   10        100  Christian
#> 40 Melbourne    2    1    7    4   11    7   15   10        100  Christian
#> 41 Melbourne    2    1    7    4   11    7   15   10        100        Tim
#> 42 Melbourne    2    1    7    4   11    7   15   10        100    Charlie
#> 43 Melbourne    2    1    7    4   11    7   15   10        100     Bernie
#> 44 Melbourne    2    1    7    4   11    7   15   10        100       Jack
#>        Surname    ID Jumper.No.      Playing.for Kicks Marks Handballs Goals
#> 1  Bontempelli 12269          4 Western Bulldogs    12     4        12     2
#> 2         Boyd 12270         17 Western Bulldogs     1     1         3     0
#> 3      Crozier 12088          9 Western Bulldogs    11     3         6     0
#> 4     Dahlhaus 11974          6 Western Bulldogs     6     4         9     0
#> 5       Daniel 12373         35 Western Bulldogs     7     3         9     0
#> 6      Dickson 12043         29 Western Bulldogs     2     1         0     0
#> 7       Gowers 12593         26 Western Bulldogs     7     1         9     1
#> 8  Honeychurch 12293         22 Western Bulldogs     7     6         7     1
#> 9       Hunter 12203          7 Western Bulldogs    20     4        12     0
#> 10  Johannisen 12122         39 Western Bulldogs     9     4        15     0
#> 11    Lipinski 12578         27 Western Bulldogs     8     2         9     0
#> 12      Macrae 12166         11 Western Bulldogs    14     3        17     0
#> 13      McLean 12363         16 Western Bulldogs    11     2        13     0
#> 14      Morris  4084         38 Western Bulldogs     2     3        10     0
#> 15    Richards 12605         20 Western Bulldogs    11     2         5     0
#> 16    Roughead 11833         23 Western Bulldogs     4     5         8     1
#> 17     Schache 12416         13 Western Bulldogs     8     2         2     1
#> 18       Smith 12390         37 Western Bulldogs     5     0         8     0
#> 19    Suckling 11718          1 Western Bulldogs    15     3        11     1
#> 20    Trengove 11812          8 Western Bulldogs    10     6         7     0
#> 21    Williams 12441         34 Western Bulldogs    15     4         8     0
#> 22        Wood 11788         10 Western Bulldogs    10     3        11     0
#> 23    Brayshaw 12322         10        Melbourne    21     8        18     1
#> 24     Fritsch 12598         31        Melbourne    14    14        12     1
#> 25        Gawn 11966         11        Melbourne     8     7        11     0
#> 26      Hannan 12492         19        Melbourne     9     5         4     1
#> 27      Harmes 12379          4        Melbourne    11     4         6     0
#> 28     Hibberd 11938         14        Melbourne    13     3        11     0
#> 29       Hogan 12321          1        Melbourne    15    10         4     5
#> 30       Jetta 11723         39        Melbourne    14     9        10     0
#> 31       Jones  2447          2        Melbourne     8     2        18     1
#> 32       Lever 12351          8        Melbourne     2     1         1     0
#> 33       Lewis  4077          6        Melbourne    11     2        17     0
#> 34    McDonald 12394         28        Melbourne     9     6         4     0
#> 35    McDonald 12009         25        Melbourne    14    11        10     4
#> 36    Melksham 11820         18        Melbourne    11     4         4     0
#> 37 Neal-Bullen 12367         30        Melbourne    12     5        13     1
#> 38      Oliver 12410         13        Melbourne    10     5        28     0
#> 39    Petracca 12430          5        Melbourne    10     4        11     1
#> 40       Salem 12272          3        Melbourne    14     5        11     0
#> 41       Smith 12514         38        Melbourne     3     2         4     0
#> 42      Spargo 12622          9        Melbourne     5     5         6     0
#> 43       Vince  4151         23        Melbourne     7     3         8     0
#> 44       Viney 12147          7        Melbourne     9     4        13     0
#>    Behinds Hit.Outs Tackles Rebounds Inside.50s Clearances Clangers Frees.For
#> 1        0        1       3        0          2          5        3         1
#> 2        0       13       2        2          1          1        2         1
#> 3        0        0       2        5          3          0        1         3
#> 4        0        0       4        1          3          3        2         1
#> 5        0        0       3        2          1          0        1         0
#> 6        0        0       2        0          1          0        0         0
#> 7        1        0       3        1          3          3        5         1
#> 8        1        0       4        1          2          2        4         1
#> 9        0        0       3        4          3          4        1         1
#> 10       1        0       1        3          3          6        7         3
#> 11       1        0       2        1          1          2        2         1
#> 12       1        0       5        5          3          5        1         4
#> 13       1        0       6        1          4          6        6         2
#> 14       0        0       2        3          0          1        1         0
#> 15       0        0       1        2          1          3        4         1
#> 16       0       12       4        1          2          2        2         0
#> 17       1        0       3        0          3          0        1         3
#> 18       0        0       2        1          1          0        2         1
#> 19       0        0       1        4          5          0        0         1
#> 20       0        0       2        4          2          0        2         0
#> 21       0        0       4        2          6          4        2         0
#> 22       0        0       2        5          1          1        3         1
#> 23       0        0       1        2          9          2        7         0
#> 24       0        0       4        7          3          2        3         2
#> 25       1       57       5        1          3          0        3         2
#> 26       1        0       1        0          5          0        2         1
#> 27       0        0       4        3          2          0        0         1
#> 28       0        0       4        4          1          0        6         2
#> 29       1        0       2        0          4          0        0         1
#> 30       0        0       3        5          2          0        1         2
#> 31       0        0       2        0          1          3        6         0
#> 32       0        0       0        1          0          0        1         0
#> 33       0        0       0        4          3          2        5         1
#> 34       0        0       1        4          1          0        3         1
#> 35       2        0       5        0          5          1        0         0
#> 36       2        0       1        0          3          2        4         0
#> 37       2        0       5        1          1          3        1         3
#> 38       1        0       3        2          6          5        3         0
#> 39       0        0       3        0          5          1        3         0
#> 40       0        0       2        6          0          0        2         0
#> 41       0        2       6        0          2          0        3         0
#> 42       0        0       2        0          4          0        2         0
#> 43       0        0       1        2          2          0        3         0
#> 44       0        0       6        1          6          6        6         2
#>    Frees.Against Brownlow.Votes Contested.Possessions Uncontested.Possessions
#> 1              1              0                    12                      12
#> 2              1              0                     4                       1
#> 3              1              0                     7                      12
#> 4              0              0                     8                       7
#> 5              0              0                     2                      14
#> 6              0              0                     0                       2
#> 7              3              0                     9                       7
#> 8              4              0                     3                      11
#> 9              0              0                     9                      20
#> 10             2              0                    12                      12
#> 11             0              0                     5                      12
#> 12             1              0                    11                      20
#> 13             1              0                    10                      14
#> 14             0              0                     5                       8
#> 15             0              0                     8                       7
#> 16             2              0                     4                       8
#> 17             0              0                     3                       7
#> 18             1              0                     6                       8
#> 19             0              0                     4                      20
#> 20             0              0                     3                      11
#> 21             0              0                     5                      18
#> 22             1              0                     7                      11
#> 23             0              2                     7                      32
#> 24             2              0                     8                      20
#> 25             2              1                    10                      11
#> 26             2              0                     2                      11
#> 27             0              0                     8                       9
#> 28             1              0                     4                      21
#> 29             0              0                     4                      14
#> 30             1              0                    11                      13
#> 31             3              0                     9                      17
#> 32             1              0                     2                       1
#> 33             0              0                     6                      21
#> 34             1              0                     6                       8
#> 35             0              3                    13                      12
#> 36             1              0                     6                       8
#> 37             0              0                     8                      18
#> 38             3              0                    19                      21
#> 39             1              0                    11                       9
#> 40             1              0                     5                      19
#> 41             1              0                     2                       5
#> 42             2              0                     1                      10
#> 43             1              0                     6                       9
#> 44             3              0                     9                      13
#>    Contested.Marks Marks.Inside.50 One.Percenters Bounces Goal.Assists
#> 1                1               0              2       0            2
#> 2                0               0              4       0            1
#> 3                0               0              2       1            0
#> 4                1               0              1       0            0
#> 5                0               0              1       0            0
#> 6                0               0              0       0            0
#> 7                0               0              0       0            0
#> 8                0               2              0       0            0
#> 9                0               0              1       3            0
#> 10               0               0              6       1            0
#> 11               0               2              0       0            0
#> 12               0               1              1       0            0
#> 13               0               0              1       0            0
#> 14               0               0              7       0            0
#> 15               0               0              2       1            0
#> 16               1               2              5       0            1
#> 17               0               1              1       0            0
#> 18               0               0              4       0            0
#> 19               0               0              0       3            1
#> 20               0               0              6       0            0
#> 21               0               0              2       0            0
#> 22               1               0              3       0            0
#> 23               1               0              1       0            2
#> 24               1               0              1       2            0
#> 25               4               3              4       0            2
#> 26               0               2              1       0            1
#> 27               2               0              2       0            1
#> 28               0               0              2       0            0
#> 29               2               5              0       0            1
#> 30               4               0              2       0            0
#> 31               0               1              0       0            0
#> 32               0               0              2       0            0
#> 33               0               0              1       0            0
#> 34               1               0              8       0            0
#> 35               6               4              4       0            1
#> 36               0               1              1       0            1
#> 37               0               4              1       0            0
#> 38               0               0              0       0            1
#> 39               1               0              0       0            2
#> 40               0               0              2       1            0
#> 41               0               0              1       0            0
#> 42               0               0              1       0            1
#> 43               0               0              0       0            0
#> 44               0               0              0       0            0
#>    Time.on.Ground..      Umpire.1 Umpire.2         Umpire.3 Umpire.4
#> 1                84 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 2                75 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 3                84 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 4                79 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 5                83 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 6                15 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 7                82 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 8                89 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 9                91 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 10               88 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 11               80 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 12               92 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 13               85 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 14               94 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 15               73 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 16               86 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 17               86 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 18               80 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 19               83 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 20               97 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 21               86 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 22               87 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 23               87 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 24               85 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 25               89 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 26               88 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 27               82 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 28               88 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 29               88 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 30               87 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 31               82 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 32               17 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 33               97 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 34              100 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 35               90 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 36               74 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 37               90 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 38               81 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 39               71 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 40               87 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 41               76 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 42               76 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 43               86 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>
#> 44               78 Brent Wallace  Sam Hay  Ray Chamberlain     <NA>

scrape_afltables_match(get_afltables_urls("01/06/2018", "01/07/2018"))

#> Downloading data

#> Error in open.connection(x, "rb"): Timeout was reached: [afltables.com] Connection timed out after 10000 milliseconds

# }