fitzRoy: Easily Scrape and Process AFL Data

An easy package for scraping and processing Australia Rules Football (AFL) data. 'fitzRoy' provides a range of functions for accessing publicly available data from 'AFL Tables' https://afltables.com/afl/afl_index.html, 'Footy Wire' https://www.footywire.com and 'The Squiggle' https://squiggle.com.au. Further functions allow for easy processing, cleaning and transformation of this data into formats that can be used for analysis.

See also

Author

Maintainer: James Day jamesthomasday@gmail.com

Authors:

Robert Nguyen nguyen.n.robert@gmail.com

Oscar Lane lane.oscar@gmail.com

Other contributors:

Matthew Erbs matthew.erbs@merbla.com [contributor]

Jason Zivkovic jaseziv83@gmail.com [contributor]

Jacob Holden jlholden26@gmail.com [contributor]