Return match URLs for specified dates

get_afltables_urls returns a character vector containing match URLs for the specified date range

get_afltables_urls(start_date, end_date = Sys.Date())

Arguments

start_date
character string for start date return to URLs from, in "dmy" or "ymd" format

end_date
optional, character string for end date to return URLS, in "dmy" or "ymd" format

Value

a character vector of match URL's between start_date and end_date

Details

This function returns match URLs for the specified date range. This will typically be used to pass to to scrape_afltables_match to return player match results.

Examples

get_afltables_urls("01/01/2018")

#>   [1] "https://afltables.com/afl/stats/games/2018/031420180322.html"
#>   [2] "https://afltables.com/afl/stats/games/2018/010520180323.html"
#>   [3] "https://afltables.com/afl/stats/games/2018/151920180324.html"
#>   [4] "https://afltables.com/afl/stats/games/2018/081320180324.html"
#>   [5] "https://afltables.com/afl/stats/games/2018/122020180324.html"
#>   [6] "https://afltables.com/afl/stats/games/2018/041020180324.html"
#>   [7] "https://afltables.com/afl/stats/games/2018/072120180325.html"
#>   [8] "https://afltables.com/afl/stats/games/2018/091120180325.html"
#>   [9] "https://afltables.com/afl/stats/games/2018/161820180325.html"
#>  [10] "https://afltables.com/afl/stats/games/2018/011420180329.html"
#>  [11] "https://afltables.com/afl/stats/games/2018/121520180330.html"
#>  [12] "https://afltables.com/afl/stats/games/2018/032020180331.html"
#>  [13] "https://afltables.com/afl/stats/games/2018/042120180331.html"
#>  [14] "https://afltables.com/afl/stats/games/2018/111920180331.html"
#>  [15] "https://afltables.com/afl/stats/games/2018/050820180331.html"
#>  [16] "https://afltables.com/afl/stats/games/2018/071820180401.html"
#>  [17] "https://afltables.com/afl/stats/games/2018/131620180401.html"
#>  [18] "https://afltables.com/afl/stats/games/2018/091020180402.html"
#>  [19] "https://afltables.com/afl/stats/games/2018/030420180406.html"
#>  [20] "https://afltables.com/afl/stats/games/2018/131920180407.html"
#>  [21] "https://afltables.com/afl/stats/games/2018/111220180407.html"
#>  [22] "https://afltables.com/afl/stats/games/2018/082020180407.html"
#>  [23] "https://afltables.com/afl/stats/games/2018/011520180407.html"
#>  [24] "https://afltables.com/afl/stats/games/2018/162120180407.html"
#>  [25] "https://afltables.com/afl/stats/games/2018/101420180408.html"
#>  [26] "https://afltables.com/afl/stats/games/2018/050720180408.html"
#>  [27] "https://afltables.com/afl/stats/games/2018/091820180408.html"
#>  [28] "https://afltables.com/afl/stats/games/2018/010420180413.html"
#>  [29] "https://afltables.com/afl/stats/games/2018/082120180414.html"
#>  [30] "https://afltables.com/afl/stats/games/2018/141920180414.html"
#>  [31] "https://afltables.com/afl/stats/games/2018/071620180414.html"
#>  [32] "https://afltables.com/afl/stats/games/2018/031220180414.html"
#>  [33] "https://afltables.com/afl/stats/games/2018/182020180414.html"
#>  [34] "https://afltables.com/afl/stats/games/2018/051320180415.html"
#>  [35] "https://afltables.com/afl/stats/games/2018/101120180415.html"
#>  [36] "https://afltables.com/afl/stats/games/2018/091520180415.html"
#>  [37] "https://afltables.com/afl/stats/games/2018/011620180420.html"
#>  [38] "https://afltables.com/afl/stats/games/2018/152120180421.html"
#>  [39] "https://afltables.com/afl/stats/games/2018/031820180421.html"
#>  [40] "https://afltables.com/afl/stats/games/2018/091320180421.html"
#>  [41] "https://afltables.com/afl/stats/games/2018/070820180421.html"
#>  [42] "https://afltables.com/afl/stats/games/2018/101220180422.html"
#>  [43] "https://afltables.com/afl/stats/games/2018/192020180422.html"
#>  [44] "https://afltables.com/afl/stats/games/2018/111420180424.html"
#>  [45] "https://afltables.com/afl/stats/games/2018/040520180425.html"
#>  [46] "https://afltables.com/afl/stats/games/2018/030720180427.html"
#>  [47] "https://afltables.com/afl/stats/games/2018/091620180428.html"
#>  [48] "https://afltables.com/afl/stats/games/2018/121320180428.html"
#>  [49] "https://afltables.com/afl/stats/games/2018/192120180428.html"
#>  [50] "https://afltables.com/afl/stats/games/2018/101520180428.html"
#>  [51] "https://afltables.com/afl/stats/games/2018/012020180428.html"
#>  [52] "https://afltables.com/afl/stats/games/2018/051120180429.html"
#>  [53] "https://afltables.com/afl/stats/games/2018/041420180429.html"
#>  [54] "https://afltables.com/afl/stats/games/2018/081820180429.html"
#>  [55] "https://afltables.com/afl/stats/games/2018/092120180504.html"
#>  [56] "https://afltables.com/afl/stats/games/2018/072020180505.html"
#>  [57] "https://afltables.com/afl/stats/games/2018/051020180505.html"
#>  [58] "https://afltables.com/afl/stats/games/2018/131820180505.html"
#>  [59] "https://afltables.com/afl/stats/games/2018/121620180505.html"
#>  [60] "https://afltables.com/afl/stats/games/2018/010320180505.html"
#>  [61] "https://afltables.com/afl/stats/games/2018/081420180506.html"
#>  [62] "https://afltables.com/afl/stats/games/2018/111520180506.html"
#>  [63] "https://afltables.com/afl/stats/games/2018/041920180506.html"
#>  [64] "https://afltables.com/afl/stats/games/2018/101620180511.html"
#>  [65] "https://afltables.com/afl/stats/games/2018/182120180512.html"
#>  [66] "https://afltables.com/afl/stats/games/2018/030520180512.html"
#>  [67] "https://afltables.com/afl/stats/games/2018/112020180512.html"
#>  [68] "https://afltables.com/afl/stats/games/2018/011320180512.html"
#>  [69] "https://afltables.com/afl/stats/games/2018/071920180512.html"
#>  [70] "https://afltables.com/afl/stats/games/2018/081520180512.html"
#>  [71] "https://afltables.com/afl/stats/games/2018/121420180513.html"
#>  [72] "https://afltables.com/afl/stats/games/2018/040920180513.html"
#>  [73] "https://afltables.com/afl/stats/games/2018/010720180518.html"
#>  [74] "https://afltables.com/afl/stats/games/2018/122120180519.html"
#>  [75] "https://afltables.com/afl/stats/games/2018/132020180519.html"
#>  [76] "https://afltables.com/afl/stats/games/2018/050920180519.html"
#>  [77] "https://afltables.com/afl/stats/games/2018/041520180519.html"
#>  [78] "https://afltables.com/afl/stats/games/2018/081620180519.html"
#>  [79] "https://afltables.com/afl/stats/games/2018/031120180520.html"
#>  [80] "https://afltables.com/afl/stats/games/2018/101920180520.html"
#>  [81] "https://afltables.com/afl/stats/games/2018/141820180520.html"
#>  [82] "https://afltables.com/afl/stats/games/2018/040720180525.html"
#>  [83] "https://afltables.com/afl/stats/games/2018/141520180526.html"
#>  [84] "https://afltables.com/afl/stats/games/2018/161920180526.html"
#>  [85] "https://afltables.com/afl/stats/games/2018/030920180526.html"
#>  [86] "https://afltables.com/afl/stats/games/2018/052120180526.html"
#>  [87] "https://afltables.com/afl/stats/games/2018/101820180527.html"
#>  [88] "https://afltables.com/afl/stats/games/2018/011120180527.html"
#>  [89] "https://afltables.com/afl/stats/games/2018/081220180527.html"
#>  [90] "https://afltables.com/afl/stats/games/2018/031620180601.html"
#>  [91] "https://afltables.com/afl/stats/games/2018/071120180602.html"
#>  [92] "https://afltables.com/afl/stats/games/2018/101320180602.html"
#>  [93] "https://afltables.com/afl/stats/games/2018/092020180602.html"
#>  [94] "https://afltables.com/afl/stats/games/2018/051420180602.html"
#>  [95] "https://afltables.com/afl/stats/games/2018/151820180602.html"
#>  [96] "https://afltables.com/afl/stats/games/2018/121920180603.html"
#>  [97] "https://afltables.com/afl/stats/games/2018/040820180603.html"
#>  [98] "https://afltables.com/afl/stats/games/2018/012120180603.html"
#>  [99] "https://afltables.com/afl/stats/games/2018/131420180608.html"
#> [100] "https://afltables.com/afl/stats/games/2018/091220180609.html"
#> [101] "https://afltables.com/afl/stats/games/2018/202120180609.html"
#> [102] "https://afltables.com/afl/stats/games/2018/151620180609.html"
#> [103] "https://afltables.com/afl/stats/games/2018/051920180610.html"
#> [104] "https://afltables.com/afl/stats/games/2018/010820180610.html"
#> [105] "https://afltables.com/afl/stats/games/2018/041120180611.html"
#> [106] "https://afltables.com/afl/stats/games/2018/071320180614.html"
#> [107] "https://afltables.com/afl/stats/games/2018/161820180615.html"
#> [108] "https://afltables.com/afl/stats/games/2018/030820180616.html"
#> [109] "https://afltables.com/afl/stats/games/2018/152020180616.html"
#> [110] "https://afltables.com/afl/stats/games/2018/011020180616.html"
#> [111] "https://afltables.com/afl/stats/games/2018/091420180617.html"
#> [112] "https://afltables.com/afl/stats/games/2018/051820180621.html"
#> [113] "https://afltables.com/afl/stats/games/2018/111320180622.html"
#> [114] "https://afltables.com/afl/stats/games/2018/102020180623.html"
#> [115] "https://afltables.com/afl/stats/games/2018/192120180623.html"
#> [116] "https://afltables.com/afl/stats/games/2018/071220180623.html"
#> [117] "https://afltables.com/afl/stats/games/2018/030420180624.html"
#> [118] "https://afltables.com/afl/stats/games/2018/141620180628.html"
#> [119] "https://afltables.com/afl/stats/games/2018/070920180629.html"
#> [120] "https://afltables.com/afl/stats/games/2018/031320180630.html"
#> [121] "https://afltables.com/afl/stats/games/2018/011820180630.html"
#> [122] "https://afltables.com/afl/stats/games/2018/042020180630.html"
#> [123] "https://afltables.com/afl/stats/games/2018/102120180630.html"
#> [124] "https://afltables.com/afl/stats/games/2018/111520180701.html"
#> [125] "https://afltables.com/afl/stats/games/2018/051220180701.html"
#> [126] "https://afltables.com/afl/stats/games/2018/081920180701.html"
#> [127] "https://afltables.com/afl/stats/games/2018/091620180705.html"
#> [128] "https://afltables.com/afl/stats/games/2018/011420180706.html"
#> [129] "https://afltables.com/afl/stats/games/2018/031920180707.html"
#> [130] "https://afltables.com/afl/stats/games/2018/131520180707.html"
#> [131] "https://afltables.com/afl/stats/games/2018/071020180707.html"
#> [132] "https://afltables.com/afl/stats/games/2018/081120180707.html"
#> [133] "https://afltables.com/afl/stats/games/2018/122020180708.html"
#> [134] "https://afltables.com/afl/stats/games/2018/040520180708.html"
#> [135] "https://afltables.com/afl/stats/games/2018/182120180708.html"
#> [136] "https://afltables.com/afl/stats/games/2018/010920180712.html"
#> [137] "https://afltables.com/afl/stats/games/2018/031520180713.html"
#> [138] "https://afltables.com/afl/stats/games/2018/101920180714.html"
#> [139] "https://afltables.com/afl/stats/games/2018/071120180714.html"
#> [140] "https://afltables.com/afl/stats/games/2018/052020180714.html"
#> [141] "https://afltables.com/afl/stats/games/2018/142120180714.html"
#> [142] "https://afltables.com/afl/stats/games/2018/041820180715.html"
#> [143] "https://afltables.com/afl/stats/games/2018/121620180715.html"
#> [144] "https://afltables.com/afl/stats/games/2018/081320180715.html"
#> [145] "https://afltables.com/afl/stats/games/2018/141520180720.html"
#> [146] "https://afltables.com/afl/stats/games/2018/041220180721.html"
#> [147] "https://afltables.com/afl/stats/games/2018/162020180721.html"
#> [148] "https://afltables.com/afl/stats/games/2018/050820180721.html"
#> [149] "https://afltables.com/afl/stats/games/2018/011920180721.html"
#> [150] "https://afltables.com/afl/stats/games/2018/091120180721.html"
#> [151] "https://afltables.com/afl/stats/games/2018/031020180722.html"
#> [152] "https://afltables.com/afl/stats/games/2018/071820180722.html"
#> [153] "https://afltables.com/afl/stats/games/2018/132120180722.html"
#> [154] "https://afltables.com/afl/stats/games/2018/051620180727.html"
#> [155] "https://afltables.com/afl/stats/games/2018/041420180728.html"
#> [156] "https://afltables.com/afl/stats/games/2018/091920180728.html"
#> [157] "https://afltables.com/afl/stats/games/2018/152120180728.html"
#> [158] "https://afltables.com/afl/stats/games/2018/032020180728.html"
#> [159] "https://afltables.com/afl/stats/games/2018/011120180728.html"
#> [160] "https://afltables.com/afl/stats/games/2018/121820180729.html"
#> [161] "https://afltables.com/afl/stats/games/2018/071320180729.html"
#> [162] "https://afltables.com/afl/stats/games/2018/081020180729.html"
#> [163] "https://afltables.com/afl/stats/games/2018/091420180803.html"
#> [164] "https://afltables.com/afl/stats/games/2018/051020180804.html"
#> [165] "https://afltables.com/afl/stats/games/2018/121920180804.html"
#> [166] "https://afltables.com/afl/stats/games/2018/011320180804.html"
#> [167] "https://afltables.com/afl/stats/games/2018/041620180804.html"
#> [168] "https://afltables.com/afl/stats/games/2018/071520180804.html"
#> [169] "https://afltables.com/afl/stats/games/2018/032120180805.html"
#> [170] "https://afltables.com/afl/stats/games/2018/112020180805.html"
#> [171] "https://afltables.com/afl/stats/games/2018/081820180805.html"
#> [172] "https://afltables.com/afl/stats/games/2018/051520180810.html"
#> [173] "https://afltables.com/afl/stats/games/2018/091020180811.html"
#> [174] "https://afltables.com/afl/stats/games/2018/142020180811.html"
#> [175] "https://afltables.com/afl/stats/games/2018/131820180811.html"
#> [176] "https://afltables.com/afl/stats/games/2018/012120180811.html"
#> [177] "https://afltables.com/afl/stats/games/2018/041920180811.html"
#> [178] "https://afltables.com/afl/stats/games/2018/071220180812.html"
#> [179] "https://afltables.com/afl/stats/games/2018/111620180812.html"
#> [180] "https://afltables.com/afl/stats/games/2018/030820180812.html"
#> [181] "https://afltables.com/afl/stats/games/2018/051420180817.html"
#> [182] "https://afltables.com/afl/stats/games/2018/041320180818.html"
#> [183] "https://afltables.com/afl/stats/games/2018/080920180818.html"
#> [184] "https://afltables.com/afl/stats/games/2018/162120180818.html"
#> [185] "https://afltables.com/afl/stats/games/2018/101520180818.html"
#> [186] "https://afltables.com/afl/stats/games/2018/192020180818.html"
#> [187] "https://afltables.com/afl/stats/games/2018/030720180819.html"
#> [188] "https://afltables.com/afl/stats/games/2018/111820180819.html"
#> [189] "https://afltables.com/afl/stats/games/2018/011220180819.html"
#> [190] "https://afltables.com/afl/stats/games/2018/051320180824.html"
#> [191] "https://afltables.com/afl/stats/games/2018/092020180825.html"
#> [192] "https://afltables.com/afl/stats/games/2018/071420180825.html"
#> [193] "https://afltables.com/afl/stats/games/2018/040820180825.html"
#> [194] "https://afltables.com/afl/stats/games/2018/010320180825.html"
#> [195] "https://afltables.com/afl/stats/games/2018/101620180825.html"
#> [196] "https://afltables.com/afl/stats/games/2018/181920180826.html"
#> [197] "https://afltables.com/afl/stats/games/2018/112120180826.html"
#> [198] "https://afltables.com/afl/stats/games/2018/121520180826.html"
#> [199] "https://afltables.com/afl/stats/games/2018/101420180906.html"
#> [200] "https://afltables.com/afl/stats/games/2018/091120180907.html"
#> [201] "https://afltables.com/afl/stats/games/2018/162120180908.html"
#> [202] "https://afltables.com/afl/stats/games/2018/041820180908.html"
#> [203] "https://afltables.com/afl/stats/games/2018/101120180914.html"
#> [204] "https://afltables.com/afl/stats/games/2018/042120180915.html"
#> [205] "https://afltables.com/afl/stats/games/2018/041420180921.html"
#> [206] "https://afltables.com/afl/stats/games/2018/111820180922.html"
#> [207] "https://afltables.com/afl/stats/games/2018/041820180929.html"
#> [208] "https://afltables.com/afl/stats/games/2019/031420190321.html"
#> [209] "https://afltables.com/afl/stats/games/2019/040920190322.html"
#> [210] "https://afltables.com/afl/stats/games/2019/111320190323.html"
#> [211] "https://afltables.com/afl/stats/games/2019/011020190323.html"
#> [212] "https://afltables.com/afl/stats/games/2019/071620190323.html"
#> [213] "https://afltables.com/afl/stats/games/2019/181920190323.html"
#> [214] "https://afltables.com/afl/stats/games/2019/152020190324.html"
#> [215] "https://afltables.com/afl/stats/games/2019/052120190324.html"
#> [216] "https://afltables.com/afl/stats/games/2019/081220190324.html"
#> [217] "https://afltables.com/afl/stats/games/2019/041420190328.html"
#> [218] "https://afltables.com/afl/stats/games/2019/011620190329.html"
#> [219] "https://afltables.com/afl/stats/games/2019/051520190330.html"
#> [220] "https://afltables.com/afl/stats/games/2019/031320190330.html"
#> [221] "https://afltables.com/afl/stats/games/2019/091120190330.html"
#> [222] "https://afltables.com/afl/stats/games/2019/182120190330.html"
#> [223] "https://afltables.com/afl/stats/games/2019/121920190331.html"
#> [224] "https://afltables.com/afl/stats/games/2019/071020190331.html"
#> [225] "https://afltables.com/afl/stats/games/2019/082020190331.html"
#> [226] "https://afltables.com/afl/stats/games/2019/010920190404.html"
#> [227] "https://afltables.com/afl/stats/games/2019/051120190405.html"
#> [228] "https://afltables.com/afl/stats/games/2019/031620190406.html"
#> [229] "https://afltables.com/afl/stats/games/2019/142120190406.html"
#> [230] "https://afltables.com/afl/stats/games/2019/131920190406.html"
#> [231] "https://afltables.com/afl/stats/games/2019/041820190406.html"
#> [232] "https://afltables.com/afl/stats/games/2019/072020190407.html"
#> [233] "https://afltables.com/afl/stats/games/2019/101220190407.html"
#> [234] "https://afltables.com/afl/stats/games/2019/081520190407.html"
#> [235] "https://afltables.com/afl/stats/games/2019/111620190411.html"
#> [236] "https://afltables.com/afl/stats/games/2019/040720190412.html"
#> [237] "https://afltables.com/afl/stats/games/2019/092120190413.html"
#> [238] "https://afltables.com/afl/stats/games/2019/051920190413.html"
#> [239] "https://afltables.com/afl/stats/games/2019/131420190413.html"
#> [240] "https://afltables.com/afl/stats/games/2019/011220190413.html"
#> [241] "https://afltables.com/afl/stats/games/2019/081820190413.html"
#> [242] "https://afltables.com/afl/stats/games/2019/032020190414.html"
#> [243] "https://afltables.com/afl/stats/games/2019/101520190414.html"
#> [244] "https://afltables.com/afl/stats/games/2019/041920190418.html"
#> [245] "https://afltables.com/afl/stats/games/2019/051220190419.html"
#> [246] "https://afltables.com/afl/stats/games/2019/131820190419.html"
#> [247] "https://afltables.com/afl/stats/games/2019/082120190420.html"
#> [248] "https://afltables.com/afl/stats/games/2019/111520190420.html"
#> [249] "https://afltables.com/afl/stats/games/2019/141620190420.html"
#> [250] "https://afltables.com/afl/stats/games/2019/030720190421.html"
#> [251] "https://afltables.com/afl/stats/games/2019/012020190421.html"
#> [252] "https://afltables.com/afl/stats/games/2019/091020190422.html"
#> [253] "https://afltables.com/afl/stats/games/2019/111420190424.html"
#> [254] "https://afltables.com/afl/stats/games/2019/040520190425.html"
#> [255] "https://afltables.com/afl/stats/games/2019/121320190426.html"
#> [256] "https://afltables.com/afl/stats/games/2019/192020190427.html"
#> [257] "https://afltables.com/afl/stats/games/2019/011520190427.html"
#> [258] "https://afltables.com/afl/stats/games/2019/162120190427.html"
#> [259] "https://afltables.com/afl/stats/games/2019/070820190427.html"
#> [260] "https://afltables.com/afl/stats/games/2019/031020190428.html"
#> [261] "https://afltables.com/afl/stats/games/2019/091820190428.html"
#> [262] "https://afltables.com/afl/stats/games/2019/041320190503.html"
#> [263] "https://afltables.com/afl/stats/games/2019/101120190504.html"
#> [264] "https://afltables.com/afl/stats/games/2019/152120190504.html"
#> [265] "https://afltables.com/afl/stats/games/2019/161920190504.html"
#> [266] "https://afltables.com/afl/stats/games/2019/071420190504.html"
#> [267] "https://afltables.com/afl/stats/games/2019/182020190504.html"
#> [268] "https://afltables.com/afl/stats/games/2019/031220190505.html"
#> [269] "https://afltables.com/afl/stats/games/2019/050920190505.html"
#> [270] "https://afltables.com/afl/stats/games/2019/010820190505.html"
#> [271] "https://afltables.com/afl/stats/games/2019/051620190510.html"
#> [272] "https://afltables.com/afl/stats/games/2019/030420190511.html"
#> [273] "https://afltables.com/afl/stats/games/2019/071920190511.html"
#> [274] "https://afltables.com/afl/stats/games/2019/112020190511.html"
#> [275] "https://afltables.com/afl/stats/games/2019/151820190511.html"
#> [276] "https://afltables.com/afl/stats/games/2019/011320190511.html"
#> [277] "https://afltables.com/afl/stats/games/2019/091220190512.html"
#> [278] "https://afltables.com/afl/stats/games/2019/102120190512.html"
#> [279] "https://afltables.com/afl/stats/games/2019/081420190512.html"
#> [280] "https://afltables.com/afl/stats/games/2019/111820190517.html"
#> [281] "https://afltables.com/afl/stats/games/2019/041520190518.html"
#> [282] "https://afltables.com/afl/stats/games/2019/011920190518.html"
#> [283] "https://afltables.com/afl/stats/games/2019/070920190518.html"
#> [284] "https://afltables.com/afl/stats/games/2019/050820190518.html"
#> [285] "https://afltables.com/afl/stats/games/2019/121620190518.html"
#> [286] "https://afltables.com/afl/stats/games/2019/132020190519.html"
#> [287] "https://afltables.com/afl/stats/games/2019/101420190519.html"
#> [288] "https://afltables.com/afl/stats/games/2019/032120190519.html"
#> [289] "https://afltables.com/afl/stats/games/2019/041620190524.html"
#> [290] "https://afltables.com/afl/stats/games/2019/101320190525.html"
#> [291] "https://afltables.com/afl/stats/games/2019/071220190525.html"
#> [292] "https://afltables.com/afl/stats/games/2019/011820190525.html"
#> [293] "https://afltables.com/afl/stats/games/2019/051420190525.html"
#> [294] "https://afltables.com/afl/stats/games/2019/092020190525.html"
#> [295] "https://afltables.com/afl/stats/games/2019/112120190526.html"
#> [296] "https://afltables.com/afl/stats/games/2019/031520190526.html"
#> [297] "https://afltables.com/afl/stats/games/2019/081920190526.html"
#> [298] "https://afltables.com/afl/stats/games/2019/121420190531.html"
#> [299] "https://afltables.com/afl/stats/games/2019/040820190601.html"
#> [300] "https://afltables.com/afl/stats/games/2019/202120190601.html"
#> [301] "https://afltables.com/afl/stats/games/2019/091620190601.html"
#> [302] "https://afltables.com/afl/stats/games/2019/101920190601.html"
#> [303] "https://afltables.com/afl/stats/games/2019/011120190601.html"
#> [304] "https://afltables.com/afl/stats/games/2019/131520190602.html"
#> [305] "https://afltables.com/afl/stats/games/2019/030520190602.html"
#> [306] "https://afltables.com/afl/stats/games/2019/071820190602.html"
#> [307] "https://afltables.com/afl/stats/games/2019/091420190607.html"
#> [308] "https://afltables.com/afl/stats/games/2019/031920190608.html"
#> [309] "https://afltables.com/afl/stats/games/2019/122020190608.html"
#> [310] "https://afltables.com/afl/stats/games/2019/012120190608.html"
#> [311] "https://afltables.com/afl/stats/games/2019/161820190609.html"
#> [312] "https://afltables.com/afl/stats/games/2019/041120190610.html"
#> [313] "https://afltables.com/afl/stats/games/2019/011420190613.html"
#> [314] "https://afltables.com/afl/stats/games/2019/051020190614.html"
#> [315] "https://afltables.com/afl/stats/games/2019/152020190615.html"
#> [316] "https://afltables.com/afl/stats/games/2019/081320190615.html"
#> [317] "https://afltables.com/afl/stats/games/2019/030720190615.html"
#> [318] "https://afltables.com/afl/stats/games/2019/122120190616.html"
#> [319] "https://afltables.com/afl/stats/games/2019/051820190620.html"
#> [320] "https://afltables.com/afl/stats/games/2019/101620190621.html"
#> [321] "https://afltables.com/afl/stats/games/2019/081120190622.html"
#> [322] "https://afltables.com/afl/stats/games/2019/151920190622.html"
#> [323] "https://afltables.com/afl/stats/games/2019/091320190622.html"
#> [324] "https://afltables.com/afl/stats/games/2019/040720190623.html"
#> [325] "https://afltables.com/afl/stats/games/2019/052120190627.html"
#> [326] "https://afltables.com/afl/stats/games/2019/010920190628.html"
#> [327] "https://afltables.com/afl/stats/games/2019/101820190629.html"
#> [328] "https://afltables.com/afl/stats/games/2019/162020190629.html"
#> [329] "https://afltables.com/afl/stats/games/2019/041220190629.html"
#> [330] "https://afltables.com/afl/stats/games/2019/071320190629.html"
#> [331] "https://afltables.com/afl/stats/games/2019/141520190630.html"
#> [332] "https://afltables.com/afl/stats/games/2019/111920190630.html"
#> [333] "https://afltables.com/afl/stats/games/2019/030820190630.html"
#> [334] "https://afltables.com/afl/stats/games/2019/041020190705.html"
#> [335] "https://afltables.com/afl/stats/games/2019/051620190706.html"
#> [336] "https://afltables.com/afl/stats/games/2019/142020190706.html"
#> [337] "https://afltables.com/afl/stats/games/2019/011320190706.html"
#> [338] "https://afltables.com/afl/stats/games/2019/070920190706.html"
#> [339] "https://afltables.com/afl/stats/games/2019/081820190706.html"
#> [340] "https://afltables.com/afl/stats/games/2019/031120190707.html"
#> [341] "https://afltables.com/afl/stats/games/2019/121520190707.html"
#> [342] "https://afltables.com/afl/stats/games/2019/192120190707.html"
#> [343] "https://afltables.com/afl/stats/games/2019/041820190712.html"
#> [344] "https://afltables.com/afl/stats/games/2019/031620190713.html"
#> [345] "https://afltables.com/afl/stats/games/2019/081020190713.html"
#> [346] "https://afltables.com/afl/stats/games/2019/051220190713.html"
#> [347] "https://afltables.com/afl/stats/games/2019/012020190713.html"
#> [348] "https://afltables.com/afl/stats/games/2019/091520190713.html"
#> [349] "https://afltables.com/afl/stats/games/2019/142120190714.html"
#> [350] "https://afltables.com/afl/stats/games/2019/071120190714.html"
#> [351] "https://afltables.com/afl/stats/games/2019/131920190714.html"
#> [352] "https://afltables.com/afl/stats/games/2019/010520190719.html"
#> [353] "https://afltables.com/afl/stats/games/2019/131420190720.html"
#> [354] "https://afltables.com/afl/stats/games/2019/032020190720.html"
#> [355] "https://afltables.com/afl/stats/games/2019/042120190720.html"
#> [356] "https://afltables.com/afl/stats/games/2019/121920190720.html"
#> [357] "https://afltables.com/afl/stats/games/2019/081620190720.html"
#> [358] "https://afltables.com/afl/stats/games/2019/091020190721.html"
#> [359] "https://afltables.com/afl/stats/games/2019/111820190721.html"
#> [360] "https://afltables.com/afl/stats/games/2019/071520190721.html"
#> [361] "https://afltables.com/afl/stats/games/2019/041420190726.html"
#> [362] "https://afltables.com/afl/stats/games/2019/101920190727.html"
#> [363] "https://afltables.com/afl/stats/games/2019/010320190727.html"
#> [364] "https://afltables.com/afl/stats/games/2019/121820190727.html"
#> [365] "https://afltables.com/afl/stats/games/2019/111520190727.html"
#> [366] "https://afltables.com/afl/stats/games/2019/132120190727.html"
#> [367] "https://afltables.com/afl/stats/games/2019/070820190728.html"
#> [368] "https://afltables.com/afl/stats/games/2019/091620190728.html"
#> [369] "https://afltables.com/afl/stats/games/2019/052020190728.html"
#> [370] "https://afltables.com/afl/stats/games/2019/101220190802.html"
#> [371] "https://afltables.com/afl/stats/games/2019/051320190803.html"
#> [372] "https://afltables.com/afl/stats/games/2019/162120190803.html"
#> [373] "https://afltables.com/afl/stats/games/2019/080920190803.html"
#> [374] "https://afltables.com/afl/stats/games/2019/111420190803.html"
#> [375] "https://afltables.com/afl/stats/games/2019/011520190803.html"
#> [376] "https://afltables.com/afl/stats/games/2019/042020190804.html"
#> [377] "https://afltables.com/afl/stats/games/2019/031820190804.html"
#> [378] "https://afltables.com/afl/stats/games/2019/071920190804.html"
#> [379] "https://afltables.com/afl/stats/games/2019/102120190809.html"
#> [380] "https://afltables.com/afl/stats/games/2019/041120190810.html"
#> [381] "https://afltables.com/afl/stats/games/2019/131620190810.html"
#> [382] "https://afltables.com/afl/stats/games/2019/192020190810.html"
#> [383] "https://afltables.com/afl/stats/games/2019/050720190810.html"
#> [384] "https://afltables.com/afl/stats/games/2019/091220190810.html"
#> [385] "https://afltables.com/afl/stats/games/2019/081520190811.html"
#> [386] "https://afltables.com/afl/stats/games/2019/031420190811.html"
#> [387] "https://afltables.com/afl/stats/games/2019/011820190811.html"
#> [388] "https://afltables.com/afl/stats/games/2019/111620190816.html"
#> [389] "https://afltables.com/afl/stats/games/2019/031520190817.html"
#> [390] "https://afltables.com/afl/stats/games/2019/091920190817.html"
#> [391] "https://afltables.com/afl/stats/games/2019/010420190817.html"
#> [392] "https://afltables.com/afl/stats/games/2019/121320190817.html"
#> [393] "https://afltables.com/afl/stats/games/2019/050820190817.html"
#> [394] "https://afltables.com/afl/stats/games/2019/141820190818.html"
#> [395] "https://afltables.com/afl/stats/games/2019/072120190818.html"
#> [396] "https://afltables.com/afl/stats/games/2019/102020190818.html"
#> [397] "https://afltables.com/afl/stats/games/2019/040520190823.html"
#> [398] "https://afltables.com/afl/stats/games/2019/151620190824.html"
#> [399] "https://afltables.com/afl/stats/games/2019/111220190824.html"
#> [400] "https://afltables.com/afl/stats/games/2019/030920190824.html"
#> [401] "https://afltables.com/afl/stats/games/2019/202120190824.html"
#> [402] "https://afltables.com/afl/stats/games/2019/101820190824.html"
#> [403] "https://afltables.com/afl/stats/games/2019/010720190825.html"
#> [404] "https://afltables.com/afl/stats/games/2019/141920190825.html"
#> [405] "https://afltables.com/afl/stats/games/2019/081320190825.html"
#> [406] "https://afltables.com/afl/stats/games/2019/051820190905.html"
#> [407] "https://afltables.com/afl/stats/games/2019/040920190906.html"
#> [408] "https://afltables.com/afl/stats/games/2019/072120190907.html"
#> [409] "https://afltables.com/afl/stats/games/2019/141920190907.html"
#> [410] "https://afltables.com/afl/stats/games/2019/091820190913.html"
#> [411] "https://afltables.com/afl/stats/games/2019/192120190914.html"
#> [412] "https://afltables.com/afl/stats/games/2019/091420190920.html"
#> [413] "https://afltables.com/afl/stats/games/2019/042120190921.html"
#> [414] "https://afltables.com/afl/stats/games/2019/142120190928.html"

get_afltables_urls("01/01/2018", end_date = "01/04/2018")

#>  [1] "https://afltables.com/afl/stats/games/2018/031420180322.html"
#>  [2] "https://afltables.com/afl/stats/games/2018/010520180323.html"
#>  [3] "https://afltables.com/afl/stats/games/2018/151920180324.html"
#>  [4] "https://afltables.com/afl/stats/games/2018/081320180324.html"
#>  [5] "https://afltables.com/afl/stats/games/2018/122020180324.html"
#>  [6] "https://afltables.com/afl/stats/games/2018/041020180324.html"
#>  [7] "https://afltables.com/afl/stats/games/2018/072120180325.html"
#>  [8] "https://afltables.com/afl/stats/games/2018/091120180325.html"
#>  [9] "https://afltables.com/afl/stats/games/2018/161820180325.html"
#> [10] "https://afltables.com/afl/stats/games/2018/011420180329.html"
#> [11] "https://afltables.com/afl/stats/games/2018/121520180330.html"
#> [12] "https://afltables.com/afl/stats/games/2018/032020180331.html"
#> [13] "https://afltables.com/afl/stats/games/2018/042120180331.html"
#> [14] "https://afltables.com/afl/stats/games/2018/111920180331.html"
#> [15] "https://afltables.com/afl/stats/games/2018/050820180331.html"