Return match URLs for specified dates

get_afltables_urls returns a character vector containing match URLs for the specified date range

get_afltables_urls(start_date, end_date = Sys.Date())

Arguments

start_date
character string for start date return to URLs from, in "dmy" or "ymd" format

end_date
optional, character string for end date to return URLS, in "dmy" or "ymd" format

Value

a character vector of match URL's between start_date and end_date

Details

This function returns match URLs for the specified date range. This will typically be used to pass to to scrape_afltables_match to return player match results.

Examples

get_afltables_urls("01/01/2018", end_date = "01/04/2018")

#>  [1] "https://afltables.com/afl/stats/games/2018/031420180322.html"
#>  [2] "https://afltables.com/afl/stats/games/2018/010520180323.html"
#>  [3] "https://afltables.com/afl/stats/games/2018/151920180324.html"
#>  [4] "https://afltables.com/afl/stats/games/2018/081320180324.html"
#>  [5] "https://afltables.com/afl/stats/games/2018/122020180324.html"
#>  [6] "https://afltables.com/afl/stats/games/2018/041020180324.html"
#>  [7] "https://afltables.com/afl/stats/games/2018/072120180325.html"
#>  [8] "https://afltables.com/afl/stats/games/2018/091120180325.html"
#>  [9] "https://afltables.com/afl/stats/games/2018/161820180325.html"
#> [10] "https://afltables.com/afl/stats/games/2018/011420180329.html"
#> [11] "https://afltables.com/afl/stats/games/2018/121520180330.html"
#> [12] "https://afltables.com/afl/stats/games/2018/032020180331.html"
#> [13] "https://afltables.com/afl/stats/games/2018/042120180331.html"
#> [14] "https://afltables.com/afl/stats/games/2018/111920180331.html"
#> [15] "https://afltables.com/afl/stats/games/2018/050820180331.html"