Get basic match results from afltables.com

get_match_results returns a dataframe containing all match results from 1897-current

get_match_results()

Value

Returns a data frame containing a line for each match

Details

The dataframe contains information about the Date, teams involved, scores and venue. It comes from afltables 'big lists' section. This is a limited dataset but is very fast to access. It generally is updated on the day after the last game

Examples

# NOT RUN {
get_match_results()
# }