Recreate the ladder for every or any given round and/or season

return_ladder returns a dataframe containing the ladder for either all seasons and rounds since 1987, or individual rounds/seasons

return_ladder(match_results_df = NA, season_round = NA, season = NA)

Arguments

match_results_df
A dataframe that has been returned from get_match_results. If empty get_match_results will execute first

season_round
An integer of the round or vector of integers for multiple rounds. If empty, all rounds returned

season
An integer of the season or vector of integers for multiple seasons. If empty, all seasons returned

Value

Returns a data frame containing a line for each team's ladder position at each round of a season

Details

The dataframe contains information about the Round, Season, Points For/Against, Ladder Position. It can either take in a data frame created using get_match_results, or if match_results_df is unspecified, will extract all games using get_match_results. Will only allow selecting rounds of the premiership season, not finals.

Examples

# \donttest{
return_ladder()

#> # A tibble: 30,358 x 8
#>    Season Team  Round.Number Season.Points Score.For Score.Against Percentage
#>     <dbl> <chr>        <int>         <dbl>     <dbl>         <dbl>      <dbl>
#>  1   1897 Fitz…            1             4        49            16      3.06 
#>  2   1897 Coll…            1             4        41            16      2.56 
#>  3   1897 Esse…            1             4        47            24      1.96 
#>  4   1897 Melb…            1             4        44            27      1.63 
#>  5   1897 Sydn…            1             0        27            44      0.614
#>  6   1897 Geel…            1             0        24            47      0.511
#>  7   1897 St K…            1             0        16            41      0.390
#>  8   1897 Carl…            1             0        16            49      0.327
#>  9   1897 Fitz…            2             8       115            42      2.74 
#> 10   1897 Melb…            2             8       108            46      2.35 
#> # … with 30,348 more rows, and 1 more variable: Ladder.Position <int>

return_ladder(match_results_df=get_match_results_df, season_round=23, season=1990:2019)

#> Error in withCallingHandlers(expr, warning = function(w) invokeRestart("muffleWarning")): object 'get_match_results_df' not found

return_ladder(season_round = 10, season = 2019)

#> # A tibble: 18 x 8
#>    Season Team  Round.Number Season.Points Score.For Score.Against Percentage
#>     <dbl> <chr>        <int>         <dbl>     <dbl>         <dbl>      <dbl>
#>  1   2019 Geel…           10            36      1003           688      1.46 
#>  2   2019 Coll…           10            32       931           730      1.28 
#>  3   2019 GWS             10            28       954           734      1.30 
#>  4   2019 Rich…           10            28       843           794      1.06 
#>  5   2019 West…           10            28       766           754      1.02 
#>  6   2019 Bris…           10            24       897           870      1.03 
#>  7   2019 Adel…           10            20       797           713      1.12 
#>  8   2019 Frem…           10            20       766           700      1.09 
#>  9   2019 Hawt…           10            20       797           768      1.04 
#> 10   2019 Port…           10            20       815           791      1.03 
#> 11   2019 St K…           10            20       743           810      0.917
#> 12   2019 Esse…           10            16       768           785      0.978
#> 13   2019 Foot…           10            16       816           868      0.940
#> 14   2019 Nort…           10            12       805           931      0.865
#> 15   2019 Sydn…           10            12       752           880      0.855
#> 16   2019 Melb…           10            12       693           914      0.758
#> 17   2019 Gold…           10            12       617           819      0.753
#> 18   2019 Carl…           10             4       705           919      0.767
#> # … with 1 more variable: Ladder.Position <int>

# }